requirements.txt
flask>=3.1.1
SQLAlchemy>=2.0.41
Flask-Session>=0.8.0
flask-sqlalchemy>=3.1.1
requests>=2.32.4
beautifulsoup4>=4.13.4
flask-talisman>=1.1.0
python-dotenv>=1.1.1
